requirements.txt
streamlit
pyspellchecker
nltk
